fix: 修复编译问题 — RDNA4 opcode 类型 + ggml BLAS 可选化
- amd_rdna4.h: RDNA4InstInfo.opcode uint8_t→uint16_t
  VOP3 指令 opcode 超过 255 (如 0x1C0), uint8_t 导致 narrowing 错误
- ggml/CMakeLists.txt: ggml_fused 改为 find_package(BLAS) 可选
  缺少 OpenBLAS 时跳过融合算子库, 不影响核心 sass_core

构建: CMake 100% 通过, 22/22 测试 PASS

+++ src/sass/backends/amd_rdna4.h
@@ -64,7 +64,7 @@ enum class AmdOpClass : uint8_t {
 struct RDNA4InstInfo {
     const char* mnemonic;      // AMD GPU 汇编助记符
     AmdEncFamily family;       // 编码家族
-    uint8_t  opcode;           // 家族内操作码
+    uint16_t opcode;           // 家族内操作码 (VOP3 可达 0x1C0+)
     uint8_t  num_src;          // 源操作数数量
     uint8_t  num_dst;          // 目标操作数数量
     uint16_t latency;          // 延迟 (wavefront 周期)
